Added room 6
Room 6 is not complete: a jump pad should launch it up to the red coin,
so it can then "catch" itself on a red platform underneath. Can't get
the coin to do it appropriately.

I sent out an email asking someone to take a look, my attempts just
didn't get it to work right.